moved utils inwards, renamed other to rest_utils, added course_exceptions
