+++ Closure_Project/rest_api/rest_utils.py
@@ -7,11 +7,19 @@ from functools import lru_cache
 from typing import Optional
 
 from project_settings.settings import AUTH0_DOMAIN, SPA_CLIENT_ID
+from .course_exceptions import COURSE_TYPE_EXCEPTIONS
 
 
 def get_course_type(track: Track, course: Course) -> Optional[CourseType]:
     if not track:
         return None
+
+    try:
+        # handles courses that are mis-categorized in the data
+        return COURSE_TYPE_EXCEPTIONS[course.course_id][track.track_number]
+    except KeyError:
+        # course isn't exceptional
+        pass
 
     if course.is_corner_stone:
         return CourseType.CORNER_STONE
@@ -32,9 +40,11 @@ def jwt_get_username_from_payload_handler(payload):
     authenticate(remote_user=username)
     return username
 
+
 @lru_cache(maxsize=1)
 def get_jwks():
     return requests.get('https://{}/.well-known/jwks.json'.format(AUTH0_DOMAIN)).json()
+
 
 def jwt_decode_token(token):
     header = jwt.get_unverified_header(token)
