--- utils.py
@@ -1,18 +1,0 @@
-import json
-import os
-
-
-def dump_json(o: object, filename: str) -> None:
-    with open(filename, 'w', encoding='utf8') as f:
-        json.dump(o, f, ensure_ascii=False)
-
-
-def load_json(filename: str):
-    with open(filename, 'r', encoding='utf8') as f:
-        return json.load(f)
-
-
-def setup_django_pycharm():
-    os.environ.setdefault("DJANGO_SETTINGS_MODULE", "project_settings.settings")
-    import django
-    django.setup()
